refactor: move drawAnt into Ant class
